Simplify component: remove props, use withClass RSC-680
