+++ gallery/src/components/NxDivider/NxDividerHorizontalExample.tsx
@@ -4,102 +4,32 @@
  * the terms of the Eclipse Public License 2.0 which accompanies this
  * distribution and is available at https://www.eclipse.org/legal/epl-2.0/.
  */
-import React, { useState, FormEvent } from 'react';
-
-import {
-  NxCheckbox,
-  NxForm,
-  NxRadio,
-  NxFormGroup,
-  NxTextInput,
-  nxTextInputStateHelpers,
-  NxFieldset,
-  useToggle,
-  NxFormSelect,
-  nxFormSelectStateHelpers,
-  NxDivider
-} from '@sonatype/react-shared-components';
-
-const { initialState, userInput } = nxTextInputStateHelpers;
+import { NxP, NxDivider, NxCode, NxButton, NxButtonBar } from '@sonatype/react-shared-components';
+import React from 'react';
 
 export default function NxDividerHorizontalExample() {
-  function validator(val: string) {
-    return val.length ? null : 'Must be non-empty';
-  }
-
-  const [selectState, setSelectVal] = nxFormSelectStateHelpers.useNxFormSelectState<string>('');
-
-  function onSelectChange(evt: FormEvent<HTMLSelectElement>) {
-    setSelectVal(evt.currentTarget.value);
-  }
-
-  const [usernameState, setUsernameState] = useState(initialState('', validator)),
-      [hostnameState, setHostnameState] = useState(initialState('')),
-      [redChecked, toggleRed] = useToggle(false),
-      [blueChecked, toggleBlue] = useToggle(false),
-      [greenChecked, toggleGreen] = useToggle(false),
-      [radioColor, setRadioColor] = useState<string | null>(null),
-      [submitCount, setSubmitCount] = useState(0);
-
-  function onUsernameChange(val: string) {
-    setUsernameState(userInput(validator, val));
-  }
-
-  function onHostnameChange(val: string) {
-    setHostnameState(userInput(validator, val));
-  }
-
-  function onSubmit() {
-    alert('Form submitted');
-    setSubmitCount(submitCount + 1);
-  }
-
   return (
-    <NxForm onSubmit={onSubmit}>
-      <NxFormGroup label="Username" isRequired>
-        <NxTextInput { ...usernameState } onChange={onUsernameChange} validatable/>
-      </NxFormGroup>
-      <NxFormGroup label="Hostname">
-        <NxTextInput { ...hostnameState } onChange={onHostnameChange} className="nx-text-input--long"/>
-      </NxFormGroup>
-      <NxDivider horizontal />
-      <NxFieldset label="Colors" isRequired>
-        <NxCheckbox onChange={toggleRed} isChecked={redChecked}>Red</NxCheckbox>
-        <NxCheckbox onChange={toggleBlue} isChecked={blueChecked}>Blue</NxCheckbox>
-        <NxCheckbox onChange={toggleGreen} isChecked={greenChecked}>Green</NxCheckbox>
-      </NxFieldset>
-      <NxDivider horizontal />
-      <NxFieldset label="Primary Color" sublabel="Pick a single color">
-        <NxRadio name="color"
-                 value="red"
-                 onChange={setRadioColor}
-                 isChecked={radioColor === 'red'}>
-          Red
-        </NxRadio>
-        <NxRadio name="color"
-                 value="purple"
-                 onChange={setRadioColor}
-                 isChecked={radioColor === 'purple'}>
-          Purple
-        </NxRadio>
-        <NxRadio name="color" value="green" onChange={setRadioColor} isChecked={radioColor === 'green'}>
-          Green
-        </NxRadio>
-        <NxRadio name="color" value="blue" onChange={setRadioColor} isChecked={radioColor === 'blue'}>
-          Blue
-        </NxRadio>
-      </NxFieldset>
-      <NxDivider horizontal />
-      <NxFormGroup label="Select">
-        <NxFormSelect { ...selectState } onChange={onSelectChange}>
-          <option value="">Select an option</option>
-          <option value="option1">Option 1</option>
-          <option value="option2">Option 2</option>
-          <option value="option3">Option 3</option>
-          <option value="option4">Option 4</option>
-          <option value="option5">Option 5</option>
-        </NxFormSelect>
-      </NxFormGroup>
-    </NxForm>
+    <>
+      <NxP>This is a simple paragraph</NxP>
+      <NxDivider/>
+      <NxP>
+        This example simply demonstrates how to use the <NxCode>NxDivider</NxCode> to render
+        simple horizontal dividers.
+      </NxP>
+      <NxDivider />
+      <NxP>
+        It is a long established fact that a reader will be distracted by the readable content of a page when
+        looking at its layout. The point of using Lorem Ipsum is that it has a more-or-less normal distribution
+        of letters, as opposed to using 'Content here, content here', making it look like readable English. Many
+        desktop publishing packages and web page editors now use Lorem Ipsum as their default model text, and a
+        search for 'lorem ipsum' will uncover many web sites still in their infancy. Various versions have evolved
+        over the years, sometimes by accident, sometimes on purpose (injected humour and the like).
+      </NxP>
+      <NxDivider />
+      <NxButtonBar>
+        <NxButton variant="primary">Enter The Matrix</NxButton>
+        <NxButton variant="primary" disabled>But Is It Real?</NxButton>
+      </NxButtonBar>
+    </>
   );
 }
