+++ lib/src/components/NxDivider/NxDivider.scss
@@ -7,16 +7,7 @@
  .nx-divider {
   background-color: var(--nx-swatch-indigo-90);
   border-radius: 1px;
- }
-
- .nx-divider--horizontal {
-   border-style: none;
-   min-height: 1px;
-   min-width: 100%;
- }
-
- .nx-divider--vertical {
-   margin: 0 var(--nx-spacing-base);
-   min-height: 100%;
-   min-width: 2px;
+  border-style: none;
+  margin: 24px 0;
+  min-height: 1px;
  }
