+++ lib/src/components/NxDivider/__tests__/NxDivider.test.tsx
@@ -4,43 +4,23 @@
  * the terms of the Eclipse Public License 2.0 which accompanies this
  * distribution and is available at https://www.eclipse.org/legal/epl-2.0/.
  */
-import NxDivider from '../NxDivider';
-import { NxDividerProps } from '../types';
+import { NxDivider } from '../NxDivider';
 import { getShallowComponent } from '../../../__testutils__/enzymeUtils';
 
 describe('NxDivider', function() {
-  const minimalProps: NxDividerProps = {};
-  const getShallow = getShallowComponent(NxDivider, minimalProps);
+  const getShallow = getShallowComponent(NxDivider, {});
 
-  it('renders a default horizontal divider', function() {
+  it('renders a divider', function() {
     const wrapper = getShallow();
-    expect(wrapper).toMatchSelector('hr.nx-divider');
+    expect(wrapper).toExist();
     expect(wrapper).toContainExactlyOneMatchingElement('hr');
+    expect(wrapper).toMatchSelector('hr.nx-divider');
   });
 
   it('renders the classNames given to it', function() {
-    const extendedProps: Partial<NxDividerProps> = {
-      className: 'test-classname ufo'
-    };
-    const wrapper = getShallow(extendedProps);
-    expect(wrapper).toMatchSelector('hr.nx-divider.test-classname.ufo');
-  });
-
-  it('renders a horizontal divider if no props are provided', function() {
-    const wrapper = getShallow();
+    const wrapper = getShallow({className: 'test-classname'});
+    expect(wrapper).toExist();
     expect(wrapper).toContainExactlyOneMatchingElement('hr');
-    expect(wrapper).toMatchSelector('hr.nx-divider.nx-divider--horizontal');
-  });
-
-  it('renders a horizontal divider by specifying the horizontal prop', function() {
-    const wrapper = getShallow({horizontal: true});
-    expect(wrapper).toContainExactlyOneMatchingElement('hr');
-    expect(wrapper).toMatchSelector('hr.nx-divider.nx-divider--horizontal');
-  });
-
-  it('renders a vertical divider by specifying the vertical prop', function() {
-    const wrapper = getShallow({vertical: true});
-    expect(wrapper).toContainExactlyOneMatchingElement('div');
-    expect(wrapper).toMatchSelector('div.nx-divider.nx-divider--vertical');
+    expect(wrapper).toMatchSelector('hr.nx-divider.test-classname');
   });
 });
