+++ lib/src/index.ts
@@ -238,4 +238,4 @@ export { default as NxTransferList, Props as NxTransferListProps, DataItem as Nx
 export { default as NxStatefulTransferList, Props as NxStatefulTransferListProps }
   from './components/NxTransferList/stateful/NxStatefulTransferList';
 
-export { default as NxDivider, NxDividerProps } from './components/NxDivider/NxDivider';
+export { NxDivider } from './components/NxDivider/NxDivider';
